Replaced API Defined UI with TOML Definitions
Removed the UI instancing code in main.rs, widgets are now all being loaded from the main.toml in the example app. There is a text positioning / text boundary size issue in this commit that needs to be fixed, most likely an issue with text boundary size calculations for taffy.

There is now a 'Window' node at the top of the uitree which is implicitly the root off all widgets that are declared without specifying a parent. The reason is all Taffy nodes should be inside the coordinate system of the window area, not coordinate systems unto themselves which leads to widgets being placed at 0,0 in vulkan space regardless of what x,y value you define.

There is a new position_control option called LayoutThenManual to fix initial positioning of manual (position controlled via Vulkan shader buffer) objects. This can be cleaned up later to make a simpler API

This technically completes Milestone 4 of #10, less a couple needed fixes.