Fix: Dynamic Text and Shape Boundaries
Text objects and shapes no longer use the static size defined in TOML config for hitboxes / boundaries.  Instead they are calculated to remain the same size as their visual on-screen size. This fixes an issue where objects would block interactivity of other widgets behind them, or cause text to not display because it extended beyond the pre set boundaries.

Will require smarter boundary rules in the future for text size, for example, if text should be limited to within a certain area, do we cut off display, perform text wrapping, or not allow more text to be input?